requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
CurrencyConverter==0.17.10
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.4
forex-python==1.8
idna==3.4
inflection==0.5.1
jsonschema-specifications==2023.7.1
jsonschema==4.19.0
pip==23.1.2
requests==2.31.0
rpds-py==0.10.3
setuptools==68.0.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
